+++ learn2track/models/gru_regression.py
@@ -1,0 +1,98 @@
+from os.path import join as pjoin
+
+import numpy as np
+import theano
+import theano.tensor as T
+from collections import OrderedDict
+
+from smartlearner.interfaces import Loss
+import smartlearner.initializers as initer
+
+from learn2track.models.layers import LayerRegression
+from learn2track.models import GRU
+
+
+class GRU_Regression(GRU):
+    """ A standard GRU model with a regression layer stacked on top of it.
+    """
+    def __init__(self, input_size, hidden_sizes, output_size, **_):
+        """
+        Parameters
+        ----------
+        input_size : int
+            Number of units each element Xi in the input sequence X has.
+        hidden_sizes : int, list of int
+            Number of hidden units each GRU should have.
+        output_size : int
+            Number of units the regression layer should have.
+        """
+        super().__init__(input_size, hidden_sizes)
+        self.output_size = output_size
+        self.layer_regression = LayerRegression(self.hidden_sizes[-1], self.output_size)
+
+    def initialize(self, weights_initializer=initer.UniformInitializer(1234)):
+        super().initialize(weights_initializer)
+        self.layer_regression.initialize(weights_initializer)
+
+    @property
+    def hyperparameters(self):
+        hyperparameters = super().hyperparameters
+        hyperparameters['output_size'] = self.output_size
+        return hyperparameters
+
+    @property
+    def parameters(self):
+        return super().parameters + self.layer_regression.parameters
+
+    def _fprop(self, Xi, *args):
+        outputs = super()._fprop(Xi, *args)
+        last_layer_h = outputs[len(self.hidden_sizes)-1]
+        regression_out = self.layer_regression.fprop(last_layer_h)
+        return outputs + (regression_out,)
+
+    def get_output(self, X):
+        outputs_info_h = []
+        for hidden_size in self.hidden_sizes:
+            outputs_info_h.append(T.zeros((X.shape[0], hidden_size)))
+
+        results, updates = theano.scan(fn=self._fprop,
+                                       outputs_info=outputs_info_h + [None],
+                                       sequences=[T.transpose(X, axes=(1, 0, 2))])  # We want to scan over sequence elements, not the examples.
+
+        self.graph_updates = updates
+        # Put back the examples so they are in the first dimension.
+        self.regression_out = T.transpose(results[-1], axes=(1, 0, 2))
+        return self.regression_out
+
+    def use(self, X):
+        directions = self.get_output(X)
+        return directions
+
+
+class L2DistanceForSequences(Loss):
+    """ Computes the L2 error of the output.
+
+    Notes
+    -----
+    This loss assumes the regression target at every time step is a vector.
+    """
+    def __init__(self, normalize_output=False):
+        self.normalize_output = normalize_output
+
+    def _get_updates(self):
+        return {}  # There is no updates for L2Distance.
+
+    def _compute_losses(self, model_output):
+        mask = self.dataset.symb_mask
+
+        # regression_outputs.shape = (batch_size, seq_length, out_dim)
+        regression_outputs = model_output
+        if self.normalize_output:
+            regression_outputs /= (T.sqrt(T.sum(regression_outputs**2, axis=2, keepdims=True) + 1e-8))
+
+        # L2_errors_per_time_step.shape = (batch_size,)
+        self.L2_errors_per_time_step = T.sqrt(T.sum(((regression_outputs - self.dataset.symb_targets)**2), axis=2))
+        # avg_L2_error_per_seq.shape = (batch_size,)
+        self.avg_L2_error_per_seq = T.sum(self.L2_errors_per_time_step*mask, axis=1) / T.sum(mask, axis=1)
+
+        return self.avg_L2_error_per_seq
